Fixed all compilation issues caused due to restructuring

+++ Trading/src/main/java/com/suneesh/trading/core/NextTradeDetails.java
@@ -6,7 +6,7 @@ import lombok.Data;
 public class NextTradeDetails {
         long tradeId;
         double amount;
-        int strategyId;
+        long strategyId;
         String callOrPut ;
         long contractDuration;
         int nextStepCount;
